separated profiles into new file, profiles.js
removed trailing whitespace
